API all setup for leaderboards

+++ API/PartyGameModels/Blackjack.cs
@@ -9,6 +9,6 @@ namespace PartyGameModels
         public int Id { get; set; }
         public string UserName { get; set; }
         public int GamesId { get; set; }
-        public float WinLossRatio { get; set; }
+        public double WinLossRatio { get; set; }
     }
 }
